Move protocols to separate files. Move workers to separate files. Fix table controller logic. Add ViewModelConfigurable protocol.
