+++ xDrip/Extensions/UIView+Effects.swift
