+++ xDrip/Scenes/Entries/EntriesList/EntriesListInteractor.swift
@@ -56,61 +56,3 @@ final class EntriesListInteractor: EntriesListBusinessLogic, EntriesListDataStor
         entriesWorker = persistenceWorker
     }
 }
-
-protocol EntriesListEntryPersistenceWorker {
-    func fetchEntries() -> [AbstractEntry]
-    func deleteEntry(_ index: Int)
-}
-
-final class EntriesListCarbsPersistenceWorker: EntriesListEntryPersistenceWorker {
-    private var carbs: [CarbEntry] = []
-    
-    func fetchEntries() -> [AbstractEntry] {
-        carbs = []
-        
-        for _ in 0 ... 20 {
-            let randValue = Double.random(in: 0...100)
-            let randomTimeInterval = TimeInterval.random(in: 0 ... 1_000_000_000)
-            let date = Date(timeIntervalSince1970: randomTimeInterval)
-            
-            let entry = CarbEntry(amount: randValue,
-                                  foodType: nil,
-                                  assimilationDuration: 0.0,
-                                  date: date)
-                carbs.append(entry)
-        }
-        
-        return carbs
-    }
-
-    func deleteEntry(_ index: Int) {
-        let entry = carbs.remove(at: index)
-        
-        // add delete from database
-    }
-}
-
-final class EntriesListBolusPersistenceWorker: EntriesListEntryPersistenceWorker {
-    private var bolusEntries: [BolusEntry] = []
-    
-    func fetchEntries() -> [AbstractEntry] {
-        bolusEntries = []
-        
-        for _ in 0 ... 20 {
-            let randValue = Double.random(in: 0...100)
-            let randomTimeInterval = TimeInterval.random(in: 0 ... 1_000_000_000)
-            let date = Date(timeIntervalSince1970: randomTimeInterval)
-            
-            let entry = BolusEntry(amount: randValue, date: date)
-                bolusEntries.append(entry)
-        }
-        
-        return bolusEntries
-    }
-
-    func deleteEntry(_ index: Int) {
-        let entry = bolusEntries.remove(at: index)
-        
-        // add delete from database
-    }
-}
