+++ xDrip/Scenes/Entries/EntriesList/EntriesListPresenter.swift
@@ -37,51 +37,3 @@ final class EntriesListPresenter: EntriesListPresentationLogic {
         self.formattingWorker = formattingWorker
     }
 }
-
-protocol EntriesListFormattingWorker {
-    func formatEntries(_ entries: [AbstractEntry]) -> [EntriesList.CellViewModel]
-}
-
-final class EntriesListCarbsFormattingWorker: EntriesListFormattingWorker {
-    
-    func formatEntries(_ entries: [AbstractEntry]) -> [EntriesList.CellViewModel] {
-        var cellViewModels: [EntriesList.CellViewModel] = []
-        
-        entries.forEach { (entry) in
-            var value = ""
-
-            value = String(format: "%.02f g", (entry as! CarbEntry).amount)
-
-            let date = DateFormatter.localizedString(from: entry.date ?? Date(), dateStyle: .short, timeStyle: .short)
-
-            cellViewModels.append(
-                EntriesList.CellViewModel(value: value,
-                                     date: date)
-            )
-        }
-        
-        return cellViewModels
-    }
-}
-
-final class EntriesListBolusFormattingWorker: EntriesListFormattingWorker {
-    
-    func formatEntries(_ entries: [AbstractEntry]) -> [EntriesList.CellViewModel] {
-        var cellViewModels: [EntriesList.CellViewModel] = []
-        
-        entries.forEach { (entry) in
-            var value = ""
-
-            value = String(format: "%.02f U", (entry as! BolusEntry).amount)
-
-            let date = DateFormatter.localizedString(from: entry.date ?? Date(), dateStyle: .short, timeStyle: .short)
-
-            cellViewModels.append(
-                EntriesList.CellViewModel(value: value,
-                                     date: date)
-            )
-        }
-        
-        return cellViewModels
-    }
-}
