+++ xDrip/Scenes/Entries/EntriesList/Protocols/EntriesListEntryPersistenceWorker.swift
@@ -1,0 +1,14 @@
+//
+//  EntriesListEntryPersistenceWorker.swift
+//  xDrip
+//
+//  Created by Ivan Skoryk on 30.03.2020.
+//  Copyright © 2020 Faifly. All rights reserved.
+//
+
+import Foundation
+
+protocol EntriesListEntryPersistenceWorker {
+    func fetchEntries() -> [AbstractEntry]
+    func deleteEntry(_ index: Int)
+}
